+++ CHANGELOG.md
@@ -1143,3 +1143,4 @@ Tab switching changed from onTap: (_) => setState(() {}) which rebuilt on every 
 - Added an in-memory dictionary _last_poi_fetch for storing time and location of the latest POI fetch to compare it on subsequent POI fetches to make sure the user is not moving impossibly fast
 - Made an is_moving_too_fast_for_poi method that ends the request early if the user is moving too fast
 - Added a specific error code for this condition which is caught by the frontend to show said message to the user
+- Added a client-side check that shows the same error as the moving too fast backend check if the user has moved too far from where they were during the request to fetch the POIs. This is purely for UX as otherwise no POIs would load with no explanation